+++ lib/byebug/commands/break.rb
@@ -23,13 +23,15 @@ module Byebug
 
     def self.description
       <<-DESCRIPTION
-        b[reak] [<file>:]<line> [if <expr>]
+        b[reak] [[<file>:]<line> [if <expr>]]
         b[reak] [<module>::...]<class>(.|#)<method> [if <expr>]
 
         #{short_description}
 
         They can be specified by line or method and an expression can be added
         for conditionally enabled breakpoints.
+        for conditionally enabled breakpoints. Without arguments create a
+        a breakpoint in the current line.
       DESCRIPTION
     end
 
@@ -38,9 +40,9 @@ module Byebug
     end
 
     def execute
-      return puts(help) unless @match[1]
+      b = line_breakpoint(frame.line.to_s) unless @match[1]
 
-      b = line_breakpoint(@match[1]) || method_breakpoint(@match[1])
+      b ||= line_breakpoint(@match[1]) || method_breakpoint(@match[1])
       return errmsg(pr("break.errors.location")) unless b
 
       return puts(pr("break.created", id: b.id, file: b.source, line: b.pos)) if syntax_valid?(@match[2])